python/requirements.txt
accelerate==0.33.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.1.0
audioread==3.0.1
Babel==2.15.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
black==24.8.0
bleach==6.1.0
boto3==1.34.153
botocore==1.34.153
bpy==4.0.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
Cython==3.0.11
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.29.2
einops==0.8.0
encodec==0.1.1
executing==2.0.1
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.15.4
fqdn==1.5.1
fsspec==2024.6.1
funcy==2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
lxml_html_clean==0.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
newspaper4k==0.9.3.1
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
Send2Trash==1.8.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
triton==3.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
zipp==3.19.2
zstandard==0.23.0
